+++ google/resource_container_cluster.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/hashicorp/errwrap"
 	"github.com/hashicorp/go-version"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
@@ -419,6 +420,11 @@ func resourceContainerCluster() *schema.Resource {
 					},
 				},
 			},
+
+			"remove_default_node_pool": {
+				Type:     schema.TypeBool,
+				Optional: true,
+			},
 		},
 	}
 }
@@ -613,6 +619,17 @@ func resourceContainerClusterCreate(d *schema.ResourceData, meta interface{}) er
 	}
 
 	log.Printf("[INFO] GKE cluster %s has been created", clusterName)
+
+	if d.Get("remove_default_node_pool").(bool) {
+		op, err := config.clientContainer.Projects.Zones.Clusters.NodePools.Delete(project, zoneName, clusterName, "default-pool").Do()
+		if err != nil {
+			return errwrap.Wrapf("Error deleting default node pool: {{err}}", err)
+		}
+		err = containerSharedOperationWait(config, op, project, zoneName, "removing default node pool", timeoutInMinutes, 3)
+		if err != nil {
+			return errwrap.Wrapf("Error deleting default node pool: {{err}}", err)
+		}
+	}
 
 	return resourceContainerClusterRead(d, meta)
 }
@@ -1129,6 +1146,17 @@ func resourceContainerClusterUpdate(d *schema.ResourceData, meta interface{}) er
 		log.Printf("[INFO] GKE cluster %s pod security policy config has been updated", d.Id())
 
 		d.SetPartial("pod_security_policy_config")
+	}
+
+	if d.HasChange("remove_default_node_pool") && d.Get("remove_default_node_pool").(bool) {
+		op, err := config.clientContainer.Projects.Zones.Clusters.NodePools.Delete(project, zoneName, clusterName, "default-pool").Do()
+		if err != nil {
+			return errwrap.Wrapf("Error deleting default node pool: {{err}}", err)
+		}
+		err = containerSharedOperationWait(config, op, project, zoneName, "removing default node pool", timeoutInMinutes, 3)
+		if err != nil {
+			return errwrap.Wrapf("Error deleting default node pool: {{err}}", err)
+		}
 	}
 
 	d.Partial(false)