+++ CHANGELOG.md
@@ -7,6 +7,7 @@ IMPROVEMENTS:
 * compute: Add `wait_for_instances` field to `google_compute_instance_group_manager` and self_link option to the `google_compute_instance_group` data source [GH-1222]
 * compute: add support for security policies in backend services [GH-1243]
 * container: add ability to delete the default node pool [GH-1245]
+* container: Add update support for pod security policy [GH-1195]
 
 BUG FIXES:
 * compute: fix panic when setting empty iap block [GH-1232]